requirements.txt
coverage==5.1
flask-restful==0.3.6
flask-script==2.0.6
ipython==7.14.0
portray==1.3.1
pre-commit==2.3.0
pytest==5.4.1
pytest-mock==3.1.0
python-logstash-async==1.4.1
requests==2.19.1
